backend/NLP_chatbot/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-generativeai
Pillow
python-dotenv
langdetect
logging